# Optimize Log2FloorNonZero with Lzcnt (.NET Core 3.0) intrinsic