--- a/BrotliSharpLib/BrotliSharpLib.csproj
+++ b/BrotliSharpLib/BrotliSharpLib.csproj
@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFrameworks>net20;net35;net40;net45;net451;net452;net462;netstandard1.3;netstandard1.6;netstandard2.0;netcoreapp1.0;netcoreapp1.1</TargetFrameworks>
+		<TargetFrameworks>net20;net35;net40;net45;net451;net452;net462;netstandard1.3;netstandard1.6;netstandard2.0;netcoreapp1.0;netcoreapp1.1;netcoreapp3.0</TargetFrameworks>
 		<AllowUnsafeBlocks>True</AllowUnsafeBlocks>
 		<GenerateDocumentationFile>true</GenerateDocumentationFile>
 		<NoWarn>1570;1701</NoWarn>