Tested on a 2mb txt file.
Compressing, level=11:
before: **0.95sec**
after:       **0.83sec**
^ netcoreapp3.0
net462: **1.23sec**

For reference, .NET Core's BrotliStream (with native lib inside): **0.48sec**

Found this bottle neck via dotTrace:
![image](https://user-images.githubusercontent.com/523221/50399984-538fd480-0794-11e9-8847-a1cfe5fc24e3.png)

PS: original impl also uses it - https://github.com/google/brotli/blob/ee2a5e1540cbd6ef883a897499d9596307f7f7f9/c/enc/fast_log.h#L25